Rename utils folder to utilities in src-react to match the other utilities folder in src
